requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Deep learning utilities
einops>=0.6.0
transformers>=4.30.0

# Graph neural networks
torch-geometric>=2.3.0

# Data handling
pyarrow>=12.0.0
fastparquet>=2023.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment tracking (optional)
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code quality
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# API server (optional)
fastapi>=0.100.0
uvicorn>=0.22.0

# Foundation model backends (optional - install separately)
# uni2ts>=0.1.0  # For Moirai
# chronos-forecasting>=0.1.0  # For Chronos
# timesfm>=0.1.0  # For TimesFM

# Explainability
shap>=0.42.0
captum>=0.6.0

# Progress bars
tqdm>=4.65.0

# HTTP requests (for data download)
requests>=2.31.0
aiohttp>=3.8.0
